Optimize hero form loading for faster LCP
